rpn 5.12.0
Reorganized things a bit and added the "estimate" operators.   Using
rpn, I determined that the mass-energy equivalence of a nanometer of
matter is the power contained in approximately 5 AA battery, but I don't
have the intuition for how big a nanogram is.  How does it compare to a
bacterium for instance?   So I added a feature to express a mass, length
or volume in common terms.   The tables need a lot of filling in, but
that will be pretty easy and fun.   I'll probably add more unit types as
well.  Energy would be a good one.