rpn 5.14.0
Multiunit conversion is now complete, as described in the previous
commit comment.

Added 'hms' as a shortcut for '[ hour minute second ]' and 'dms' as a
shortcut for '[ degree arcminute arcsecond ]'.

It works like a peach, but there are cumulative errors that really bug
me.  I wouldn't expect that to happen, but apparently it does.  See the
help example for the 'hms' operator for an example of the error.