porting smart factorization code
There's a temporary new operator called 'factor2', which uses advanced
algorithms that I'm porting to Python.   When it's complete, I'll get
rid of the old 'factor' code and rename 'factor2' to 'factor'.

It's not done yet, but it works for some values.  rpn can now factor the
first 100 fibonacci numbers in less than a second, which may be faster
than it can pull them out of the cache.

The first 150 fibonaccis can be factored in less than 6 seconds.