started implementing decorated callers
This is going to take a long time to finish and eventually touch almost
every function in the whole project.  Right now, I'm deploying the
oneArgFunctionEvaluator decorator to all operators that take one
argument.

I also have to eliminate any lambdas in the operator table, although I
don't know yet if that means the lambdas needed for generators as well.

I am also not really sure how I'm going to handle operators with more
than 2 arguments.  Right now, we don't handle list arguments as well as
we do for one or two argument operators.  But the evaluator for 3+
argument operators that does the same thing could be prohibitively
complex.  Well, maybe not.  But I will definitely feel too lazy to write
it.